flask_model_qas/requirements.txt
Babel==2.9.1
camelot-py==0.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
coloredlogs==15.0
configobj==5.0.6
configparser==5.0.2
cryptography==3.4.7
decorator==4.4.2
et-xmlfile==1.1.0
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
Flask-WTF==0.14.3
future==0.18.2
httplib2==0.19.1
humanfriendly==9.1
icu==0.0.1
idna==2.10
img2pdf==0.4.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
langdetect==1.0.9
langid==1.1.6
lxml==4.6.3
MarkupSafe==1.1.1
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.0
nltk==3.6.2
numpy==1.20.2
ocrmypdf==12.0.1
opencv-contrib-python==3.4.13.47
opencv-python==4.5.2.52
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pathlib==1.0.1
pdf2image==1.14.0
pdfminer.six==20200517
pdfplumber==0.5.27
pikepdf==2.11.4
Pillow==8.2.0
pluggy==0.13.1
prov==2.0.0
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
pymongo==3.11.4
PyMuPDF==1.18.13
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyreadline==2.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyxnat==1.4
rdflib==5.0.0
regex==2021.4.4
reportlab==3.5.67
requests==2.25.1
scipy==1.6.3
simplejson==3.17.2
six==1.16.0
sortedcontainers==2.3.0
tqdm==4.60.0
traits==6.2.0
urllib3==1.26.4
Wand==0.6.6
Werkzeug==1.0.1
WTForms==2.3.3
